+++ argopy/tests/test_utilities.py
@@ -1,5 +1,4 @@
 import os
-import io
 import pytest
 import tempfile
 import xarray as xr
@@ -12,12 +11,9 @@ from argopy.utilities import (
     linear_interpolation_remap,
     format_oneline,
     wmo2box,
-    modified_environ,
     wrap_longitude,
     toYearFraction, YearFraction_to_datetime,
     argo_split_path,
-    Registry,
-    float_wmo,
     get_coriolis_profile_id,
     get_ea_profile_page,
 )
@@ -35,182 +31,5 @@ from utils import (
 from mocked_http import mocked_httpserver, mocked_server_address
 
 
-@pytest.mark.parametrize("conda", [False, True],
-                         indirect=False,
-                         ids=["conda=%s" % str(p) for p in [False, True]])
-def test_show_versions(conda):
-    f = io.StringIO()
-    argopy.show_versions(file=f, conda=conda)
-    assert "SYSTEM" in f.getvalue()
-
-
-class Test_linear_interpolation_remap:
-    @pytest.fixture(autouse=True)
-    def create_data(self):
-        # create fake data to test interpolation:
-        temp = np.random.rand(200, 100)
-        pres = np.sort(
-            np.floor(
-                np.zeros([200, 100])
-                + np.linspace(50, 950, 100)
-                + np.random.randint(-5, 5, [200, 100])
-            )
-        )
-        self.dsfake = xr.Dataset(
-            {
-                "TEMP": (["N_PROF", "N_LEVELS"], temp),
-                "PRES": (["N_PROF", "N_LEVELS"], pres),
-            },
-            coords={
-                "N_PROF": ("N_PROF", range(200)),
-                "N_LEVELS": ("N_LEVELS", range(100)),
-                "Z_LEVELS": ("Z_LEVELS", np.arange(100, 900, 20)),
-            },
-        )
-
-    def test_interpolation(self):
-        # Run it with success:
-        dsi = linear_interpolation_remap(
-            self.dsfake["PRES"],
-            self.dsfake["TEMP"],
-            self.dsfake["Z_LEVELS"],
-            z_dim="N_LEVELS",
-            z_regridded_dim="Z_LEVELS",
-        )
-        assert "remapped" in dsi.dims
-
-    def test_interpolation_1d(self):
-        # Run it with success:
-        dsi = linear_interpolation_remap(
-            self.dsfake["PRES"].isel(N_PROF=0),
-            self.dsfake["TEMP"].isel(N_PROF=0),
-            self.dsfake["Z_LEVELS"],
-            z_regridded_dim="Z_LEVELS",
-        )
-        assert "remapped" in dsi.dims
-
-    def test_error_zdim(self):
-        # Test error:
-        # catches error from _regular_interp linked to z_dim
-        with pytest.raises(RuntimeError):
-            linear_interpolation_remap(
-                self.dsfake["PRES"],
-                self.dsfake["TEMP"],
-                self.dsfake["Z_LEVELS"],
-                z_regridded_dim="Z_LEVELS",
-            )
-
-    def test_error_ds(self):
-        # Test error:
-        # catches error from linear_interpolation_remap linked to datatype
-        with pytest.raises(ValueError):
-            linear_interpolation_remap(
-                self.dsfake["PRES"],
-                self.dsfake,
-                self.dsfake["Z_LEVELS"],
-                z_dim="N_LEVELS",
-                z_regridded_dim="Z_LEVELS",
-            )
-
-
-def test_format_oneline():
-    s = "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore"
-    assert isinstance(format_oneline(s), str)
-    assert isinstance(format_oneline(s[0:5]), str)
-    s = format_oneline(s, max_width=12)
-    assert isinstance(s, str) and len(s) == 12
-
-
-def test_modified_environ():
-    os.environ["DUMMY_ENV_ARGOPY"] = 'initial'
-    with modified_environ(DUMMY_ENV_ARGOPY='toto'):
-        assert os.environ['DUMMY_ENV_ARGOPY'] == 'toto'
-    assert os.environ['DUMMY_ENV_ARGOPY'] == 'initial'
-    os.environ.pop('DUMMY_ENV_ARGOPY')
-
-
-def test_wmo2box():
-    with pytest.raises(ValueError):
-        wmo2box(12)
-    with pytest.raises(ValueError):
-        wmo2box(8000)
-    with pytest.raises(ValueError):
-        wmo2box(2000)
-
-    def complete_box(b):
-        b2 = b.copy()
-        b2.insert(4, 0.)
-        b2.insert(5, 10000.)
-        return b2
-
-    assert is_box(complete_box(wmo2box(1212)))
-    assert is_box(complete_box(wmo2box(3324)))
-    assert is_box(complete_box(wmo2box(5402)))
-    assert is_box(complete_box(wmo2box(7501)))
-
-
-def test_wrap_longitude():
-    assert wrap_longitude(np.array([-20])) == 340
-    assert wrap_longitude(np.array([40])) == 40
-    assert np.all(np.equal(wrap_longitude(np.array([340, 20])), np.array([340, 380])))
-
-
-def test_toYearFraction():
-    assert toYearFraction(pd.to_datetime('202001010000')) == 2020
-    assert toYearFraction(pd.to_datetime('202001010000', utc=True)) == 2020
-    assert toYearFraction(pd.to_datetime('202001010000')+pd.offsets.DateOffset(years=1)) == 2021
-
-
-def test_YearFraction_to_datetime():
-    assert YearFraction_to_datetime(2020) == pd.to_datetime('202001010000')
-    assert YearFraction_to_datetime(2020+1) == pd.to_datetime('202101010000')
-
-
-class Test_argo_split_path:
-    #############
-    # UTILITIES #
-    #############
-    # src = "https://data-argo.ifremer.fr/dac"
-    src = argopy.tutorial.open_dataset("gdac")[0] + "/dac"
-    list_of_files = [
-        src + "/bodc/6901929/6901929_prof.nc",  # core / multi-profile
-        src + "/coriolis/3902131/3902131_Sprof.nc",  # bgc / synthetic multi-profile
-
-        src + "/meds/4901079/profiles/D4901079_110.nc",  # core / mono-profile / Delayed
-        src + "/aoml/13857/profiles/R13857_001.nc",  # core / mono-profile / Real
-
-        src + "/coriolis/3902131/profiles/SD3902131_001.nc",  # bgc / synthetic mono-profile / Delayed
-        src + "/coriolis/3902131/profiles/SD3902131_001D.nc",  # bgc / synthetic mono-profile / Delayed / Descent
-        src + "/coriolis/6903247/profiles/SR6903247_134.nc",  # bgc / synthetic mono-profile / Real
-        src + "/coriolis/6903247/profiles/SR6903247_134D.nc",  # bgc / synthetic mono-profile / Real / Descent
-
-        src + "/coriolis/3902131/profiles/BR3902131_001.nc",  # bgc / mono-profile / Real
-        src + "/coriolis/3902131/profiles/BR3902131_001D.nc",  # bgc / mono-profile / Real / Descent
-
-        src + "/aoml/5900446/5900446_Dtraj.nc",  # traj / Delayed
-        src + "/csio/2902696/2902696_Rtraj.nc",  # traj / Real
-
-        src + "/coriolis/3902131/3902131_BRtraj.nc",  # bgc / traj / Real
-        # src + "/coriolis/6903247/6903247_BRtraj.nc",  # bgc / traj / Real
-
-        src + "/incois/2902269/2902269_tech.nc",  # technical
-        # src + "/nmdis/2901623/2901623_tech.nc",  # technical
-
-        src + "/jma/4902252/4902252_meta.nc",  # meta-data
-        # src + "/coriolis/1900857/1900857_meta.nc",  # meta-data
-    ]
-    list_of_files = [f.replace("/", os.path.sep) for f in list_of_files]
-
-    #########
-    # TESTS #
-    #########
-
-    @pytest.mark.parametrize("file", list_of_files,
-                             indirect=False)
-    def test_argo_split_path(self, file):
-        desc = argo_split_path(file)
-        assert isinstance(desc, dict)
-        for key in ['origin', 'path', 'name', 'type', 'extension', 'wmo', 'dac']:
-            assert key in desc
