+++ argopy/tests/test_utils_compute.py
@@ -1,0 +1,75 @@
+import pytest
+import numpy as np
+import xarray as xr
+
+from argopy.utils.compute import linear_interpolation_remap
+
+
+class Test_linear_interpolation_remap:
+    @pytest.fixture(autouse=True)
+    def create_data(self):
+        # create fake data to test interpolation:
+        temp = np.random.rand(200, 100)
+        pres = np.sort(
+            np.floor(
+                np.zeros([200, 100])
+                + np.linspace(50, 950, 100)
+                + np.random.randint(-5, 5, [200, 100])
+            )
+        )
+        self.dsfake = xr.Dataset(
+            {
+                "TEMP": (["N_PROF", "N_LEVELS"], temp),
+                "PRES": (["N_PROF", "N_LEVELS"], pres),
+            },
+            coords={
+                "N_PROF": ("N_PROF", range(200)),
+                "N_LEVELS": ("N_LEVELS", range(100)),
+                "Z_LEVELS": ("Z_LEVELS", np.arange(100, 900, 20)),
+            },
+        )
+
+    def test_interpolation(self):
+        # Run it with success:
+        dsi = linear_interpolation_remap(
+            self.dsfake["PRES"],
+            self.dsfake["TEMP"],
+            self.dsfake["Z_LEVELS"],
+            z_dim="N_LEVELS",
+            z_regridded_dim="Z_LEVELS",
+        )
+        assert "remapped" in dsi.dims
+
+    def test_interpolation_1d(self):
+        # Run it with success:
+        dsi = linear_interpolation_remap(
+            self.dsfake["PRES"].isel(N_PROF=0),
+            self.dsfake["TEMP"].isel(N_PROF=0),
+            self.dsfake["Z_LEVELS"],
+            z_regridded_dim="Z_LEVELS",
+        )
+        assert "remapped" in dsi.dims
+
+    def test_error_zdim(self):
+        # Test error:
+        # catches error from _regular_interp linked to z_dim
+        with pytest.raises(RuntimeError):
+            linear_interpolation_remap(
+                self.dsfake["PRES"],
+                self.dsfake["TEMP"],
+                self.dsfake["Z_LEVELS"],
+                z_regridded_dim="Z_LEVELS",
+            )
+
+    def test_error_ds(self):
+        # Test error:
+        # catches error from linear_interpolation_remap linked to datatype
+        with pytest.raises(ValueError):
+            linear_interpolation_remap(
+                self.dsfake["PRES"],
+                self.dsfake,
+                self.dsfake["Z_LEVELS"],
+                z_dim="N_LEVELS",
+                z_regridded_dim="Z_LEVELS",
+            )
+
