+++ argopy/tests/test_utils_geo.py
@@ -1,0 +1,42 @@
+import pytest
+import numpy as np
+import pandas as pd
+from argopy.utils.geo import wmo2box, wrap_longitude, toYearFraction, YearFraction_to_datetime
+from argopy.utils.checkers import is_box
+
+
+def test_wmo2box():
+    with pytest.raises(ValueError):
+        wmo2box(12)
+    with pytest.raises(ValueError):
+        wmo2box(8000)
+    with pytest.raises(ValueError):
+        wmo2box(2000)
+
+    def complete_box(b):
+        b2 = b.copy()
+        b2.insert(4, 0.)
+        b2.insert(5, 10000.)
+        return b2
+
+    assert is_box(complete_box(wmo2box(1212)))
+    assert is_box(complete_box(wmo2box(3324)))
+    assert is_box(complete_box(wmo2box(5402)))
+    assert is_box(complete_box(wmo2box(7501)))
+
+
+def test_wrap_longitude():
+    assert wrap_longitude(np.array([-20])) == 340
+    assert wrap_longitude(np.array([40])) == 40
+    assert np.all(np.equal(wrap_longitude(np.array([340, 20])), np.array([340, 380])))
+
+
+def test_toYearFraction():
+    assert toYearFraction(pd.to_datetime('202001010000')) == 2020
+    assert toYearFraction(pd.to_datetime('202001010000', utc=True)) == 2020
+    assert toYearFraction(pd.to_datetime('202001010000')+pd.offsets.DateOffset(years=1)) == 2021
+
+
+def test_YearFraction_to_datetime():
+    assert YearFraction_to_datetime(2020) == pd.to_datetime('202001010000')
+    assert YearFraction_to_datetime(2020+1) == pd.to_datetime('202101010000')
